updated line 415
The issue is that if you group a DataFrame by one or more column all of whose values are None (i.e. missing), then the resulting grouped summary will be blank. Then this blank result flows all the rest of the way downstream. By replacing missingness with zero (or anything, really).